new_requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
dask==2022.2.0
Flask==2.1.1
fonttools==4.33.0
fsspec==2022.3.0
idna==3.3
imageio==2.17.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
locket==1.0.0
MarkupSafe==2.1.1
matplotlib==3.5.1
networkx==2.6.3
numpy==1.21.6
packaging==21.3
partd==1.2.0
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.7.3
six==1.16.0
tifffile==2021.11.2
toolz==0.11.2
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
